+++ demo/DemoV1a/Model/Base/L1Flags.generated.cs
@@ -8,7 +8,7 @@ using Newtonsoft.Json.Linq;
 
 namespace DemoV1a.Model.Base {
 /// AUTOGENERED BY caffoa ///
-    public partial class L1Flags {
+    public partial class L1Flags : IEquatable<L1Flags> {
         public const string L1FlagsObjectName = "flags";
         [JsonProperty("id")]
         public virtual string Id { get; set; }